+++ Code/ska_particle/Impl/ColoredRectGraphicParticleRenderer.cpp
@@ -20,8 +20,8 @@ void ska::ColoredRectGraphicParticleRenderer::display(const ParticleGroup& group
 		r.h = m_particleDimensions;
 		r.w = m_particleDimensions;
 		for (std::size_t i = 0; i < maxSize; i++) {
-			r.x = group.pos[i].x - r.w / 2;
-			r.y = group.pos[i].y - r.h / 2;
+			r.x = static_cast<int>(group.pos[i].x - r.w / 2);
+			r.y = static_cast<int>(group.pos[i].y - r.h / 2);
 			m_renderer.drawColorRect(group.color[i], r);
 		}
 	}
